% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSplineFormula.R
\name{bSplineFormula}
\alias{bSplineFormula}
\title{To a formula, add a set of terms corresponding to a set of b-spline basis functions}
\usage{
bSplineFormula(
  formIn,
  basisVar,
  groupingVar = "",
  nbasis = 5,
  nbasis_groupingVar = 0,
  dropLargest = 2,
  dropLargest_groupingVar = 0
)
}
\arguments{
\item{formIn}{Formula,}

\item{basisVar}{Character. The name of the numeric variable to construct the basis over.}

\item{groupingVar}{Character. Optional. If provided, return a string formatted for a random effects structure with no intercept.}

\item{nbasis}{Number of bases. Note that, in most cases, the returned number will be smaller than this due to \code{dropLargest}.}

\item{dropLargest}{Number of bases to drop, to improve identifiability.}
}
\description{
Construct a set of bases using \code{basisVar} and add the terms from these bases
to the formula \code{formIn}. Relies on \code{\link[fda]{create.bspline.basis}} and
\code{\link[fda]{getbasismatrix}} to construct bases.
}
\details{
Given the numeric vector \code{basisVar} and the number of bases \code{nbasis},
an attribute \code{attr(output,'basisDF')} is returned giving the density [weight]
for each of the bases at each of the observations.

To increase identifiability in common use cases, such as within conventional mixed-effects models,
by default the "largest" bases are removed (e.g., the basis with the mode at the largest
value of \code{basisVar} is deleted.). If identifiability is still posing difficulties,
\code{dropLargest} could be increased, in order to remove an increasing
number of the bases with modes nearest the largest values of \code{basisVar}. Alternatively, if the 
full coverage is desired, \code{dropLargest} should be lowered to 0 in order to not
remove any bases.
}
\examples{

c()

}
