% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSplineMat.R
\name{bSplineMat}
\alias{bSplineMat}
\title{Make a data frame of b-spline bases}
\usage{
bSplineMat(basisVar, nbasis = 5, dropLargest = 1, returnDF = F)
}
\arguments{
\item{basisVar}{Character. The name of the numeric variable to construct the basis over.}

\item{nbasis}{Number of bases. Note that, in most cases, the returned number will be smaller than this due to \code{dropLargest}.}

\item{dropLargest}{Number of bases to drop, to improve identifiability.}

\item{returnDF}{If TRUE, return a data frame rather than matrix}
}
\description{
Make a data frame of b-spline bases
}
\examples{

d <- data.frame(trial_number = 1:100)

# get a variable with a b-spline with 5 bases, but dropping the one with the mode centered on the largest trial number
d$spl5dr1 <- bSplineMat(d$trial_number)

 # get a variable with a b-spline with 9 bases, but dropping the three with the modes centered on the largest trial numbers
d$spl9dr3 <- bSplineMat(d$trial_number,nbasis = 9, dropLargest = 3)

}
