% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bSplineTerm.R
\name{bSplineTerm}
\alias{bSplineTerm}
\title{Define a b-spline term for a model formula}
\usage{
bSplineTerm(
  basisVar,
  nbasis = 5,
  groupingVar = "",
  moderators = c(),
  dropLargest = 1
)
}
\arguments{
\item{basisVar}{Character. The name of the numeric variable to construct the basis over.}

\item{nbasis}{Number of bases. Note that, in most cases, the returned number will be smaller than this due to \code{dropLargest}.}

\item{groupingVar}{Character. Optional. If provided, return a string formatted for a random effects structure with no intercept.}

\item{moderators}{Character vector. Optional. If provided, names of variables to interact with bases.}

\item{dropLargest}{Number of bases to drop, to improve identifiability.}
}
\description{
Define a b-spline term for a model formula
}
\examples{

c()

}
