% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cumulative.R
\name{fit_cumulative}
\alias{fit_cumulative}
\title{Fit a model to cumulative data}
\usage{
fit_cumulative(unimplemented)
}
\arguments{
\item{unimplemented}{}
}
\description{
Unimplemented
}
\examples{

c()

}
