% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_exGauss.R
\name{fit_exGauss}
\alias{fit_exGauss}
\title{Fit a response time model using an ex-Gaussian distribution}
\usage{
fit_exGauss(
  rtVar,
  expoFormula = expoMean ~ 1,
  dat,
  gaussFormula = gaussMean ~ 1,
  sigmaFormula = sigma ~ 1,
  priorIn = "empirical",
  ...
)
}
\arguments{
\item{rtVar}{Character. The name of the response time variable to be predicted.}

\item{expoFormula}{linear [mixed-effects] model formula defining the exponential component of the ex-Gaussian distribution.}

\item{dat}{Data frame}

\item{gaussFormula}{linear [mixed-effects] model formula defining the mean of the Gaussian component of the ex-Gaussian distribution.}

\item{sigmaFormula}{linear [mixed-effects] model formula defining the dispersion of the Gaussian component of the ex-Gaussian distribution.}

\item{priorIn}{Priors to provide to the model. For defaults see \code{Description}.}

\item{...}{}
}
\description{
Fits a 3-parameter ex-Gaussian model using \code{\link[brms]{brm}},
with some adjusted defaults in order to make things somewhat easier for the user.
}
\details{
By default, \code{brms} has the main model formula predict the sum of the exponential and
Gaussian components of the ex-Gaussian distribution. This function separates these so they
can be estimated separately from one another. This allows simpler partitioning of the model into
its constituent components. For instance, it may be theoretically meaningful to have one variable
of interest predict the exponential component and a different variable of interest predict the 
Gaussian component.

Default priors are defined as follows: The prior for the exponential component intercept,
which by \code{brms} default is estimated on a log scale, is defined as
a normal distribution with a mean of \code{mean(log(rt))} and a standard deviation of
\code{mad(log(rt))}; non-intercept coefficients have a prior of a normal distribution 
with a mean of 0 and a standard deviation of \code{mad(log(rt))}.
The prior for mean of the Gaussian component intercept is defined as
a normal distribution with a mean of \code{exp(mean(log(rt)))} and a standard deviation of
\code{mad(rt)}; non-intercept coefficients have a prior of a normal distribution 
with a mean of 0 and a standard deviation of \code{mad(rt)}.
}
\examples{

d <- TEfits::anstrain

## generate example response time data
d$absRat <- abs(d$ratio)
d$rt <- brms::rwiener(nrow(d)
,delta = .2 + d$absRat/5 + scale(d$trialNum)/20
, tau = .15
, alpha = 1 + sin(d$trialNum/20)/10
, beta = .5)[,'q']

expoFormula <- bSplineFormula(expoMean ~ (0 + scale(absRat) | subID) , basisVar = d$trialNum, groupingVar = 'subID')

m_EG <- fit_exGauss('rt'
,expoFormula = expoFormula
,gaussFormula = gaussMean ~ scale(sizeRat) + (1 | subID)
,dat = d
, cores = 2
)

}
