% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_SLN.R
\name{fit_SLN}
\alias{fit_SLN}
\title{Fit a response time model using a shifted lognormal distribution}
\usage{
fit_SLN(
  rtFormula,
  dat,
  sigmaFormula = sigma ~ 1,
  ndtFormula = scaleLogisNDT ~ 1,
  ndtBound = "auto",
  priorIn = prior(normal(0, 1.5), nlpar = "scaleLogisNDT"),
  ...
)
}
\arguments{
\item{rtFormula}{linear [mixed-effects] model formula defining model to predict RT (i.e., the mean of the lognormal distribution)}

\item{dat}{Data frame}

\item{sigmaFormula}{linear [mixed-effects] model formula defining sigma}

\item{ndtFormula}{linear [mixed-effects] model formula defining non-decision time (NDT). NDT is estimated as a scaled logistic function, which helps with sampling.}

\item{ndtBound}{The scalar numeric upper bound of the scaled logistic function defining NDT. Defaults to an upper bound that is the 5th percentile of the RT distribution (i.e., \code{quantile(rt,.05)}). If the model is not initializing successfully, reducing this value may help.}

\item{priorIn}{Priors to provide to the model. Note the default value which provides a reasonable prior over the possible values of NDT; if providing user-defined priors, it is recommended to also provide this one as well.}

\item{...}{}
}
\description{
Fits a 3-parameter shifted lognormal model using \code{\link[brms]{brm}},
with some adjusted defaults in order to make things somewhat easier for the user.
}
\examples{

d <- TEfits::anstrain

## generate example response time data
d$absRat <- abs(d$ratio)
d$rt <- brms::rwiener(nrow(d)
,delta = .2 + d$absRat/5 + scale(d$trialNum)/20
, tau = .15
, alpha = 1 + sin(d$trialNum/20)/10
, beta = .5)[,'q']

rtFormula <- bSplineFormula(rt ~ (0 + scale(absRat) | subID) , basisVar = d$trialNum, groupingVar = 'subID')

m_SLN <- fit_SLN(rtFormula = rtFormula
,ndtFormula = scaleLogisNDT ~ (1 | subID)
,dat = d
, cores = 2
)

}
