% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoVar.R
\name{monoVar}
\alias{monoVar}
\title{Create an ordered incremented variable}
\usage{
monoVar(continVar, return_numeric = T, increment_size = "auto")
}
\arguments{
\item{continVar}{Numeric vector, such as trial number.}

\item{return_numeric}{The grouped incremented numeric variable can be returned. By default, however, this is turned into an ordered factor and then returned.}

\item{increment_size}{Size of the smallest level, and the incremental increase in size with each level. Defaults to \code{ceiling(length(unique(continVar))/500)+1}, which means that the increment size is 2 if there are fewer than 500 unique values in \code{continVar}.}
}
\description{
Creates an ordered factor from a numeric vector.
 
In situations where we expect a monotonic change over time with 
most change happening fairly rapidly, we could want to treat a continuous
variable as an ordered factor. This function puts the smallest \code{increment_size} values
into a single factor level, the next \code{increment_size*2} values into the next smallest
factor levels, the next \code{increment_size*3} into the third of the factor levels, the next
eight into the fourth, and so on. The last factor level is combined such 
that the largest values do not have a factor level with smaller members
than the second-smallest factor level.
}
\details{
Note that increment sizes larger than 1 are recommended. This is because, for a common application 
(i.e., a \code{brms} model), there should be fewer than 50 total levels.
}
