% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfChar.R
\name{pfChar}
\alias{pfChar}
\title{Get a psychometric function formula in character form}
\usage{
pfChar(pfType)
}
\arguments{
\item{pfType}{The type of psychometric function. Currently supported are \code{logistic} or \code{weibull}}
}
\description{
Get a psychometric function formula in character form
}
\examples{

c()

}
