% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfFormula.R
\name{pfFormula}
\alias{pfFormula}
\title{Define a psychometric function formula}
\usage{
pfFormula(
  yVar,
  pf = c("identity", "logistic", "weibull"),
  link_x = "LINK_X_VARIABLE",
  ...
)
}
\arguments{
\item{yVar}{The variable to predict in the psychometric function; probably a binary numeric (i.e., 0 or 1).}

\item{pf}{Character. The name of the psychometric function you want (i.e., "logistic" or "weibull"), or "identity" for no psychometric function}

\item{link_x}{The linking \code{x}}

\item{...}{}
}
\description{
Define a psychometric function formula
}
\examples{

c()

}
